uslepp takes uint32_t-parameter
git-svn-id: svn+ssh://sgh.dk/hdc/svn/aos/trunk@76 01777da7-cf65-4787-a2ab-fdced703dd4c

+++ src/include/kernel.h
@@ -29,7 +29,7 @@ void block();
 void unblock(struct task_t* task);
 void yield();
 void msleep(uint16_t ms);
-void usleep(uint16_t us);
+void usleep(uint32_t us);
 
 extern struct task_t* current;
 extern uint8_t do_context_switch;
